move LocationRequired to components folder
