+++ apps/ledger-live-mobile/src/components/LocationRequired/OpenAppPermissionsSettingsButton.tsx
