+++ apps/ledger-live-mobile/src/components/LocationRequired/index.tsx
